email_service/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.3.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.28.0
gunicorn==21.2.0
kombu==5.5.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
pika==1.3.2
